pyproject.toml
[tool.poetry]
name = "similarity_search"
version = "0.1.0"
description = ""
authors = ["pdoyle5000 <pdoyle5000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.8.1"
invoke = "^1.5.0"
flake8 = "^3.9.0"
pytest = "^6.2.3"
pytest-flake8 = "^1.0.7"
mypy = "^0.812"
torchvision = "^0.9.1"
tqdm = "^4.60.0"
requests = "^2.25.1"
Pillow = "^8.2.0"
dash = "^1.20.0"
faiss-gpu = "^1.7.1"
isort = "^5.8.0"
mlflow = "^1.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
